refactor: consolidate change of basis functions into a dedicated module
- Moved change of basis functions from individual files to a new `change_basis.py` module for better organization and maintainability.
- Updated import statements across the codebase to reflect the new module structure.
- Enhanced documentation within the new module to clarify the purpose and usage of each function.
